fixing a bug with the way we passed around backingbuffers (and a test that reproduced the issue)
when we store ByteBuffers to hold onto to maybe use as a buffer to
serialize into in the future, we need to make sure we properly duplicate
the ByteBuffer instances (not the underlying byte arrays themselves, but
the ByteBuffer wrappers) so the various members (position, limit, etc.)
don't conflict with one another when they're used.

for example, if we parse an RTP packet from a single ByteBuffer, we held
that ByteBuffer in every place we parsed (e.g. the header and the
packet) so when we serialize the header as part of serializing the
packet, the header will override the limit of the same ByteBuffer the
packet is using to write into.  this commit fixes that by making sure we
duplicate ByteBuffers so each instance has its own view of just its
chunk of the backing data.