remove modifyInPlace-like functions for now
i was unable to get it to work the way i wanted(which was to be able to
pass in a reciever function) because we can't override the receiver
function type in subclasses to be the appropriate type.

also, though, at this point i think it's fundamentally flawed as we need
the hook of toImmutable to make sure all the data is correct.  maybe can
think of a way in the future