...a lot of stuff.
I was trying to keep this separate, but it just got to be too much.
this contains the update for RtcpHeaderData moving over to incremental,
but also contains changes for how we model srtp/srtcp packets,
srtp/srtcp-related fixes, and a (regretful?) move over to unsigned
types.  THERE ARE IMPORTANT FIXES HERE.  if you need to revert it, god
help you, because there are changes in here you almost certainly do not
want to revert (and potentially some you would...like the damn unsigned
types)