tweak the implementation of subBuffer to give an independent capacity value
before, the capacity value would be shared from the original buffer
which made it impossible to track how much of a given subBuffer was
really "yours" to work with (we couldn't trust capacity, as it was
unclear whether or not the caller intended to 'give' all of that space
and limit can change; in fact it's after we've shrunken limit and want
to reduce it when we want to know how much is 'ours' to expand into).
now capacity can serve as the number to denote how much space is 'given'
when creating a sub-buffer.